# Improve NSubstitute performance