--- a/src/NSubstitute/Core/Argument.cs
+++ b/src/NSubstitute/Core/Argument.cs
@@ -5,10 +5,14 @@ namespace NSubstitute.Core
 {
     public class Argument
     {
+        private readonly ICall _call;
+        private readonly int _argIndex;
+
         private readonly Type _declaredType;
         private readonly Func<object> _getValue;
         private readonly Action<object> _setValue;
 
+        [Obsolete("This constructor overload is deprecated and will be removed in the next version.")]
         public Argument(Type declaredType, Func<object> getValue, Action<object> setValue)
         {
             _declaredType = declaredType;
@@ -16,23 +20,33 @@ namespace NSubstitute.Core
             _setValue = setValue;
         }
 
+        public Argument(ICall call, int argIndex)
+        {
+            _call = call ?? throw new ArgumentNullException(nameof(call));
+            _argIndex = argIndex;
+        }
+
         public object Value
         {
-            get { return _getValue(); }
-            set { _setValue(value); }
+            get => _getValue != null ? _getValue() : _call.GetArguments()[_argIndex];
+            set
+            {
+                if (_setValue != null)
+                {
+                    _setValue(value);
+                }
+                else
+                {
+                    _call.GetArguments()[_argIndex] = value;
+                }
+            }
         }
 
-        public bool IsByRef { get { return DeclaredType.IsByRef; } }
+        public bool IsByRef => DeclaredType.IsByRef;
 
-        public Type DeclaredType
-        {
-            get { return _declaredType; }
-        }
+        public Type DeclaredType => _declaredType ?? _call.GetParameterInfos()[_argIndex].ParameterType;
 
-        public virtual Type ActualType
-        {
-            get { return (Value == null) ? _declaredType : Value.GetType(); }
-        }
+        public Type ActualType => Value == null ? DeclaredType : Value.GetType();
 
         public bool IsDeclaredTypeEqualToOrByRefVersionOf(Type type)
         {
@@ -51,7 +65,7 @@ namespace NSubstitute.Core
 
         private static Type AsNonByRefType(Type type)
         {
-            return (type.IsByRef) ? type.GetElementType() : type;
+            return type.IsByRef ? type.GetElementType() : type;
         }
     }
 }