--- a/src/NSubstitute/Core/ArgumentSpecificationDequeue.cs
+++ b/src/NSubstitute/Core/ArgumentSpecificationDequeue.cs
@@ -7,17 +7,24 @@ namespace NSubstitute.Core
 {
     public class ArgumentSpecificationDequeue : IArgumentSpecificationDequeue
     {
+        private static readonly IArgumentSpecification[] EmptySpecifications = new IArgumentSpecification[0];
+
         private readonly Func<IList<IArgumentSpecification>> _dequeueAllQueuedArgSpecs;
 
         public ArgumentSpecificationDequeue(Func<IList<IArgumentSpecification>> dequeueAllQueuedArgSpecs)
         {
             _dequeueAllQueuedArgSpecs = dequeueAllQueuedArgSpecs;
         }
-        
+
         public IList<IArgumentSpecification> DequeueAllArgumentSpecificationsForMethod(MethodInfo methodInfo)
         {
             if (methodInfo.GetParameters().Length == 0)
-                return new List<IArgumentSpecification>();
+            {
+                // We violate public contract, as mutable list was expected as result.
+                // However, in reality we never expect value to be mutated, so this optimization is fine.
+                // We are not allowed to change public contract due to SemVer, so keeping that as it is.
+                return EmptySpecifications;
+            }
 
             var queuedArgSpecifications = _dequeueAllQueuedArgSpecs.Invoke();
             return queuedArgSpecifications;