--- a/src/NSubstitute/Core/CallFactory.cs
+++ b/src/NSubstitute/Core/CallFactory.cs
@@ -10,19 +10,12 @@ namespace NSubstitute.Core
     {
         public ICall Create(MethodInfo methodInfo, object[] arguments, object target, IList<IArgumentSpecification> argumentSpecifications, Func<object> baseMethod)
         {
-            var parameterInfos = GetParameterInfoFromMethod(methodInfo);
-            return new Call(methodInfo, arguments, target, argumentSpecifications, parameterInfos, baseMethod);
+            return new Call(methodInfo, arguments, target, argumentSpecifications , baseMethod);
         }
 
         public ICall Create(MethodInfo methodInfo, object[] arguments, object target, IList<IArgumentSpecification> argumentSpecifications)
         {
-            var parameterInfos = GetParameterInfoFromMethod(methodInfo);
-            return new Call(methodInfo, arguments, target, argumentSpecifications, parameterInfos, baseMethod: null);
-        }
-
-        private static IParameterInfo[] GetParameterInfoFromMethod(MethodInfo methodInfo)
-        {
-            return methodInfo.GetParameters().Select(x => new ParameterInfoWrapper(x)).ToArray();
+            return new Call(methodInfo, arguments, target, argumentSpecifications, baseMethod: null);
         }
     }
 }