--- a/src/NSubstitute/Core/DefaultForType.cs
+++ b/src/NSubstitute/Core/DefaultForType.cs
@@ -5,6 +5,11 @@ namespace NSubstitute.Core
 {
     public class DefaultForType : IDefaultForType
     {
+        private static readonly object BoxedBoolean = default(bool);
+        private static readonly object BoxedInt = default(int);
+        private static readonly object BoxedLong = default(long);
+        private static readonly object BoxedDouble = default(double);
+
         public object GetDefaultFor(Type type)
         {
             if (IsVoid(type)) return null;
@@ -19,6 +24,24 @@ namespace NSubstitute.Core
 
         private object DefaultInstanceOfValueType(Type returnType)
         {
+            // Performance optimization for the most popular types.
+            if (returnType == typeof(bool))
+            {
+                return BoxedBoolean;
+            }
+            if (returnType == typeof(int))
+            {
+                return BoxedInt;
+            }
+            if (returnType == typeof(long))
+            {
+                return BoxedLong;
+            }
+            if (returnType == typeof(double))
+            {
+                return BoxedDouble;
+            }
+
             return Activator.CreateInstance(returnType);
         }
     }