--- a/src/NSubstitute/Core/IArgumentSpecificationDequeue.cs
+++ b/src/NSubstitute/Core/IArgumentSpecificationDequeue.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Reflection;
 using NSubstitute.Core.Arguments;
 
@@ -6,6 +7,9 @@ namespace NSubstitute.Core
 {
     public interface IArgumentSpecificationDequeue
     {
+        [Obsolete("This method is deprecated and will be removed in future versions of product.")]
         IList<IArgumentSpecification> DequeueAllArgumentSpecificationsForMethod(MethodInfo methodInfo);
+
+        IList<IArgumentSpecification> DequeueAllArgumentSpecificationsForMethod(int parametersCount);
     }
 }