… .travis.yml
… src/NSubstitute/Core/RouteFactoryCacheWrapper.cs
… src/NSubstitute/Core/ThreadLocalContext.cs
… src/NSubstitute/Routing/Handlers/PropertySetterHandler.cs
… src/NSubstitute/Routing/Handlers/ReturnAutoValue.cs
… src/NSubstitute/Routing/Handlers/ReturnFromAndConfigureDynamicCall.cs
… src/NSubstitute/Routing/Handlers/ReturnFromCustomHandlers.cs
… tests/NSubstitute.Benchmarks/ActivationBenchmark.cs
… tests/NSubstitute.Benchmarks/ArgumentSpecificationUsageBenchmark.cs
… tests/NSubstitute.Benchmarks/DispatchCallBenchmark.cs
… tests/NSubstitute.Benchmarks/DispatchConfiguredMatchingCallBenchmark.cs
… tests/NSubstitute.Benchmarks/DispatchConfiguredNonMatchingCallBenchmark.cs
… tests/NSubstitute.Benchmarks/DispatchNonConfiguredCallBenchmark.cs
… tests/NSubstitute.Benchmarks/NSubstitute.Benchmarks.csproj
… tests/NSubstitute.Benchmarks/TestTypes/AbstractClassWithSingleMethod.cs
… tests/NSubstitute.Benchmarks/TestTypes/ClassWithSingleMethod.cs
… tests/NSubstitute.Benchmarks/TestTypes/CustomDelegate.cs
… tests/NSubstitute.Benchmarks/TestTypes/Delegates.cs
… tests/NSubstitute.Benchmarks/TestTypes/IInterfaceWithSingleMethod.cs
… tests/NSubstitute.Benchmarks/ToStringCallBenchmark.cs
… tests/NSubstitute.Benchmarks/VerifyReceivedCallBenchmark.cs