I've reviewed the [mocking container performance comparison](https://github.com/stevedesmond-ca/BenchmarkMockNet) and found that unacceptable that Moq and FakeItEasy beat us with such a big distance 😟  Therefore I spent a couple of weeks and tried to make NSubstitute a bit faster. Moq is still faster, but we are not that much behind now.

I'm happy it was possible to achieve boost with relatively small amount of internal changes. Also notice, that I drastically decreased amount of memory allocations.

I extended our benchmarks to have more and cover basic usage scenarios. Here you can find the performance difference (took me a 3-4 hours to actually run all the benchmarks twice 😮).

<details>
  <summary>See benchmark results</summary>

```
|                              Method |  Branch | Runtime |       Mean |     Error |    StdDev |  Gen 0 | Allocated |
//
// Activation
//
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|       CreateAbstractClassSubstitute |  master |     Clr |   4.911 us | 0.0963 us | 0.1110 us | 1.4343 |   5.88 KB |
|       CreateAbstractClassSubstitute |      PR |     Clr |   3.948 us | 0.0188 us | 0.0176 us | 0.9842 |   4.05 KB |
|       CreateAbstractClassSubstitute |  master |    Core |   4.761 us | 0.0948 us | 0.2514 us | 2.0981 |   8.61 KB |
|       CreateAbstractClassSubstitute |      PR |    Core |   4.009 us | 0.0186 us | 0.0145 us | 1.0986 |   4.51 KB |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|            CreateDelegateSubstitute |  master |     Clr |   9.917 us | 0.1960 us | 0.1925 us | 1.5106 |   6.24 KB |
|            CreateDelegateSubstitute |      PR |     Clr |   9.522 us | 0.0887 us | 0.0741 us | 1.0681 |   4.42 KB |
|            CreateDelegateSubstitute |  master |    Core |   6.857 us | 0.0258 us | 0.0215 us | 2.1896 |   8.97 KB |
|            CreateDelegateSubstitute |      PR |    Core |   6.149 us | 0.0279 us | 0.0261 us | 1.1826 |   4.87 KB |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|           CreateInterfaceSubstitute |  master |     Clr |   4.540 us | 0.0294 us | 0.0230 us | 1.4954 |   6.14 KB |
|           CreateInterfaceSubstitute |      PR |     Clr |   4.516 us | 0.0139 us | 0.0130 us | 1.0529 |   4.32 KB |
|           CreateInterfaceSubstitute |  master |    Core |   4.811 us | 0.0189 us | 0.0176 us | 2.1667 |   8.88 KB |
|           CreateInterfaceSubstitute |      PR |    Core |   4.255 us | 0.0819 us | 0.0943 us | 1.1597 |   4.78 KB |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|    CreateNonAbstractClassSubstitute |  master |     Clr |   4.188 us | 0.0288 us | 0.0255 us | 1.4343 |   5.88 KB |
|    CreateNonAbstractClassSubstitute |      PR |     Clr |   3.893 us | 0.0420 us | 0.0351 us | 0.9842 |   4.05 KB |
|    CreateNonAbstractClassSubstitute |  master |    Core |   4.808 us | 0.0962 us | 0.0900 us | 2.0981 |   8.61 KB |
|    CreateNonAbstractClassSubstitute |      PR |    Core |   3.949 us | 0.0143 us | 0.0134 us | 1.0986 |   4.51 KB |
//
// ArgumentSpecificationUsage
//
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|       ConfigureArgumentWithAnyValue |  master |     Clr |   3.047 us | 0.0609 us | 0.0854 us | 0.2899 |    1.8 KB |
|       ConfigureArgumentWithAnyValue |      PR |     Clr |   2.002 us | 0.0477 us | 0.0446 us | 0.1678 |    1064 B |
|       ConfigureArgumentWithAnyValue |  master |    Core |   3.307 us | 0.0660 us | 0.1085 us | 0.2556 |   1.59 KB |
|       ConfigureArgumentWithAnyValue |      PR |    Core |   2.213 us | 0.0207 us | 0.0184 us | 0.1526 |     960 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|    ConfigureOutArgumentWithAnyValue |  master |     Clr |   3.202 us | 0.0716 us | 0.1026 us | 0.2899 |    1.8 KB |
|    ConfigureOutArgumentWithAnyValue |      PR |     Clr |   2.159 us | 0.0229 us | 0.0214 us | 0.1678 |    1064 B |
|    ConfigureOutArgumentWithAnyValue |  master |    Core |   3.503 us | 0.0534 us | 0.0474 us | 0.2556 |   1.59 KB |
|    ConfigureOutArgumentWithAnyValue |      PR |    Core |   2.405 us | 0.0139 us | 0.0116 us | 0.1526 |     960 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|    ConfigureRefArgumentWithAnyValue |  master |     Clr |   3.153 us | 0.0526 us | 0.0440 us | 0.2899 |    1.8 KB |
|    ConfigureRefArgumentWithAnyValue |      PR |     Clr |   2.314 us | 0.0253 us | 0.0211 us | 0.1678 |    1064 B |
|    ConfigureRefArgumentWithAnyValue |  master |    Core |   3.496 us | 0.0603 us | 0.0535 us | 0.2556 |   1.59 KB |
|    ConfigureRefArgumentWithAnyValue |      PR |    Core |   2.401 us | 0.0183 us | 0.0153 us | 0.1526 |     960 B |
//
// DispatchConfiguredMatchingCall
//
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|  DispatchAbstractClassProxyCall_Int |  master |     Clr |   2.995 us | 0.0585 us | 0.0601 us | 0.3548 |   2.18 KB |
|  DispatchAbstractClassProxyCall_Int |      PR |     Clr |   1.237 us | 0.0182 us | 0.0170 us | 0.0916 |     581 B |
|  DispatchAbstractClassProxyCall_Int |  master |    Core |   3.623 us | 0.0283 us | 0.0265 us | 0.3357 |   2.05 KB |
|  DispatchAbstractClassProxyCall_Int |      PR |    Core |   1.133 us | 0.0183 us | 0.0163 us | 0.0896 |     568 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
| DispatchAbstractClassProxyCall_Void |  master |     Clr |   3.717 us | 0.0262 us | 0.0245 us | 0.3815 |   2.34 KB |
| DispatchAbstractClassProxyCall_Void |      PR |     Clr |   2.031 us | 0.0273 us | 0.0255 us | 0.1144 |     717 B |
| DispatchAbstractClassProxyCall_Void |  master |    Core |   4.071 us | 0.0272 us | 0.0255 us | 0.3510 |   2.15 KB |
| DispatchAbstractClassProxyCall_Void |      PR |    Core |   1.996 us | 0.0183 us | 0.0171 us | 0.0992 |     696 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|   DispatchClassPartialProxyCall_Int |  master |     Clr |   4.183 us | 0.0828 us | 0.1952 us | 0.3891 |   2.42 KB |
|   DispatchClassPartialProxyCall_Int |      PR |     Clr |   2.135 us | 0.0296 us | 0.0262 us | 0.1297 |     829 B |
|   DispatchClassPartialProxyCall_Int |  master |    Core |   4.478 us | 0.0669 us | 0.0626 us | 0.3738 |    2.3 KB |
|   DispatchClassPartialProxyCall_Int |      PR |    Core |   2.052 us | 0.0401 us | 0.0394 us | 0.1297 |     824 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|  DispatchClassPartialProxyCall_Void |  master |     Clr |   4.668 us | 0.0511 us | 0.0478 us | 0.4196 |   2.58 KB |
|  DispatchClassPartialProxyCall_Void |      PR |     Clr |   3.071 us | 0.0300 us | 0.0235 us | 0.1526 |     965 B |
|  DispatchClassPartialProxyCall_Void |  master |    Core |   5.191 us | 0.0548 us | 0.0485 us | 0.3891 |   2.52 KB |
|  DispatchClassPartialProxyCall_Void |      PR |    Core |   2.996 us | 0.0322 us | 0.0252 us | 0.1411 |     952 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|            DispatchDelegateCall_Int |  master |     Clr |   2.662 us | 0.0573 us | 0.0858 us | 0.3433 |   2.11 KB |
|            DispatchDelegateCall_Int |      PR |     Clr |   0.929 us | 0.0162 us | 0.0144 us | 0.0820 |     517 B |
|            DispatchDelegateCall_Int |  master |    Core |   3.232 us | 0.0168 us | 0.0157 us | 0.3242 |   1.99 KB |
|            DispatchDelegateCall_Int |      PR |    Core |   0.899 us | 0.0070 us | 0.0059 us | 0.0801 |     520 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|           DispatchDelegateCall_Void |  master |     Clr |   3.337 us | 0.0196 us | 0.0184 us | 0.3700 |   2.28 KB |
|           DispatchDelegateCall_Void |      PR |     Clr |   1.765 us | 0.0139 us | 0.0130 us | 0.1030 |     653 B |
|           DispatchDelegateCall_Void |  master |    Core |   3.927 us | 0.0299 us | 0.0265 us | 0.3357 |   2.21 KB |
|           DispatchDelegateCall_Void |      PR |    Core |   1.656 us | 0.0179 us | 0.0159 us | 0.0896 |     600 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|      DispatchInterfaceProxyCall_Int |  master |     Clr |   2.698 us | 0.0540 us | 0.0739 us | 0.3433 |   2.11 KB |
|      DispatchInterfaceProxyCall_Int |      PR |     Clr |   0.955 us | 0.0186 us | 0.0174 us | 0.0820 |     517 B |
|      DispatchInterfaceProxyCall_Int |  master |    Core |   3.322 us | 0.0264 us | 0.0234 us | 0.3242 |   1.99 KB |
|      DispatchInterfaceProxyCall_Int |      PR |    Core |   0.918 us | 0.0602 us | 0.0047 us | 0.0801 |     520 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|     DispatchInterfaceProxyCall_Void |  master |     Clr |   3.264 us | 0.0219 us | 0.0204 us | 0.3700 |   2.28 KB |
|     DispatchInterfaceProxyCall_Void |      PR |     Clr |   1.793 us | 0.0174 us | 0.0155 us | 0.1030 |     653 B |
|     DispatchInterfaceProxyCall_Void |  master |    Core |   3.929 us | 0.0224 us | 0.0198 us | 0.3357 |   2.09 KB |
|     DispatchInterfaceProxyCall_Void |      PR |    Core |   1.702 us | 0.0218 us | 0.0204 us | 0.0896 |     600 B |
//
// DispatchConfiguredNonMatchingCall
//
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|  DispatchAbstractClassProxyCall_Int |  master |     Clr |   3.870 us | 0.0289 us | 0.0241 us | 0.4120 |   2.55 KB |
|  DispatchAbstractClassProxyCall_Int |      PR |     Clr |   1.561 us | 0.0123 us | 0.0115 us | 0.0744 |     469 B |
|  DispatchAbstractClassProxyCall_Int |  master |    Core |   4.373 us | 0.0213 us | 0.0189 us | 0.3738 |   2.44 KB |
|  DispatchAbstractClassProxyCall_Int |      PR |    Core |   1.502 us | 0.0178 us | 0.0167 us | 0.0725 |     456 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
| DispatchAbstractClassProxyCall_Void |  master |     Clr |   3.214 us | 0.0224 us | 0.0209 us | 0.2899 |   1.79 KB |
| DispatchAbstractClassProxyCall_Void |      PR |     Clr |   1.503 us | 0.0096 us | 0.0090 us | 0.0668 |     421 B |
| DispatchAbstractClassProxyCall_Void |  master |    Core |   3.476 us | 0.0173 us | 0.0153 us | 0.2708 |   1.73 KB |
| DispatchAbstractClassProxyCall_Void |      PR |    Core |   1.363 us | 0.0166 us | 0.0156 us | 0.0648 |     440 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|   DispatchClassPartialProxyCall_Int |  master |     Clr |   5.063 us | 0.0560 us | 0.0496 us | 0.4501 |   2.79 KB |
|   DispatchClassPartialProxyCall_Int |      PR |     Clr |   2.397 us | 0.0403 us | 0.0377 us | 0.1144 |     717 B |
|   DispatchClassPartialProxyCall_Int |  master |    Core |   5.510 us | 0.0318 us | 0.0282 us | 0.4120 |   2.69 KB |
|   DispatchClassPartialProxyCall_Int |      PR |    Core |   2.297 us | 0.0440 us | 0.0432 us | 0.1106 |     712 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|  DispatchClassPartialProxyCall_Void |  master |     Clr |   4.075 us | 0.0615 us | 0.0576 us | 0.3281 |   2.04 KB |
|  DispatchClassPartialProxyCall_Void |      PR |     Clr |   2.340 us | 0.0406 us | 0.0380 us | 0.1068 |     669 B |
|  DispatchClassPartialProxyCall_Void |  master |    Core |   4.262 us | 0.0814 us | 0.0871 us | 0.3128 |   1.92 KB |
|  DispatchClassPartialProxyCall_Void |      PR |    Core |   2.138 us | 0.0426 us | 0.0438 us | 0.1030 |     664 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|            DispatchDelegateCall_Int |  master |     Clr |   3.424 us | 0.0268 us | 0.0224 us | 0.4044 |   2.49 KB |
|            DispatchDelegateCall_Int |      PR |     Clr |   1.221 us | 0.0170 us | 0.0159 us | 0.0648 |     405 B |
|            DispatchDelegateCall_Int |  master |    Core |   3.963 us | 0.0108 us | 0.0085 us | 0.3662 |   2.38 KB |
|            DispatchDelegateCall_Int |      PR |    Core |   1.194 us | 0.0156 us | 0.0146 us | 0.0610 |     424 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|           DispatchDelegateCall_Void |  master |     Clr |   2.812 us | 0.0172 us | 0.0161 us | 0.2823 |   1.73 KB |
|           DispatchDelegateCall_Void |      PR |     Clr |   1.164 us | 0.0135 us | 0.0126 us | 0.0572 |     357 B |
|           DispatchDelegateCall_Void |  master |    Core |   3.226 us | 0.0290 us | 0.0257 us | 0.2594 |   1.61 KB |
|           DispatchDelegateCall_Void |      PR |    Core |   1.106 us | 0.0214 us | 0.0220 us | 0.0534 |     376 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|      DispatchInterfaceProxyCall_Int |  master |     Clr |   3.509 us | 0.0288 us | 0.0269 us | 0.4044 |   2.49 KB |
|      DispatchInterfaceProxyCall_Int |      PR |     Clr |   1.249 us | 0.0188 us | 0.0176 us | 0.0648 |     405 B |
|      DispatchInterfaceProxyCall_Int |  master |    Core |   4.048 us | 0.0252 us | 0.0210 us | 0.3662 |   2.38 KB |
|      DispatchInterfaceProxyCall_Int |      PR |    Core |   1.203 us | 0.0172 us | 0.0161 us | 0.0610 |     424 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|     DispatchInterfaceProxyCall_Void |  master |     Clr |   2.934 us | 0.0615 us | 0.1336 us | 0.2823 |   1.73 KB |
|     DispatchInterfaceProxyCall_Void |      PR |     Clr |   1.167 us | 0.0147 us | 0.0137 us | 0.0572 |     357 B |
|     DispatchInterfaceProxyCall_Void |  master |    Core |   3.167 us | 0.0196 us | 0.0174 us | 0.2594 |   1.61 KB |
|     DispatchInterfaceProxyCall_Void |      PR |    Core |   1.097 us | 0.0208 us | 0.0223 us | 0.0534 |     344 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
//
// DispatchNonConfiguredCall
//
|  DispatchAbstractClassProxyCall_Int |  master |     Clr |   3.821 us | 0.0359 us | 0.0336 us | 0.3624 |   2.23 KB |
|  DispatchAbstractClassProxyCall_Int |      PR |     Clr |   1.593 us | 0.0116 us | 0.0109 us | 0.0668 |     421 B |
|  DispatchAbstractClassProxyCall_Int |  master |    Core |   3.711 us | 0.0337 us | 0.0315 us | 0.3204 |   1.99 KB |
|  DispatchAbstractClassProxyCall_Int |      PR |    Core |   1.346 us | 0.0161 us | 0.0151 us | 0.0648 |     440 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
| DispatchAbstractClassProxyCall_Void |  master |     Clr |   2.905 us | 0.0223 us | 0.0209 us | 0.2480 |   1.53 KB |
| DispatchAbstractClassProxyCall_Void |      PR |     Clr |   1.376 us | 0.0128 us | 0.0120 us | 0.0668 |     421 B |
| DispatchAbstractClassProxyCall_Void |  master |    Core |   2.737 us | 0.0314 us | 0.0294 us | 0.2289 |   1.47 KB |
| DispatchAbstractClassProxyCall_Void |      PR |    Core |   1.287 us | 0.0144 us | 0.0135 us | 0.0648 |     440 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|   DispatchClassPartialProxyCall_Int |  master |     Clr |   4.605 us | 0.0473 us | 0.0442 us | 0.4044 |   2.47 KB |
|   DispatchClassPartialProxyCall_Int |      PR |     Clr |   2.298 us | 0.0449 us | 0.0398 us | 0.1068 |     669 B |
|   DispatchClassPartialProxyCall_Int |  master |    Core |   4.642 us | 0.0714 us | 0.0667 us | 0.3662 |   2.24 KB |
|   DispatchClassPartialProxyCall_Int |      PR |    Core |   2.084 us | 0.0401 us | 0.0430 us | 0.1030 |     664 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|  DispatchClassPartialProxyCall_Void |  master |     Clr |   3.819 us | 0.0309 us | 0.0289 us | 0.2861 |   1.77 KB |
|  DispatchClassPartialProxyCall_Void |      PR |     Clr |   2.358 us | 0.0392 us | 0.0367 us | 0.1068 |     669 B |
|  DispatchClassPartialProxyCall_Void |  master |    Core |   3.529 us | 0.0344 us | 0.0322 us | 0.2670 |   1.66 KB |
|  DispatchClassPartialProxyCall_Void |      PR |    Core |   2.053 us | 0.0398 us | 0.0426 us | 0.1030 |     664 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|            DispatchDelegateCall_Int |  master |     Clr |   3.240 us | 0.0209 us | 0.0195 us | 0.3510 |   2.17 KB |
|            DispatchDelegateCall_Int |      PR |     Clr |   1.112 us | 0.0047 us | 0.0037 us | 0.0572 |     357 B |
|            DispatchDelegateCall_Int |  master |    Core |   3.308 us | 0.0121 us | 0.0113 us | 0.3128 |   1.93 KB |
|            DispatchDelegateCall_Int |      PR |    Core |   1.069 us | 0.0210 us | 0.0225 us | 0.0534 |     344 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|           DispatchDelegateCall_Void |  master |     Clr |   2.868 us | 0.0324 us | 0.0303 us | 0.2365 |   1.47 KB |
|           DispatchDelegateCall_Void |      PR |     Clr |   1.113 us | 0.0157 us | 0.0147 us | 0.0572 |     357 B |
|           DispatchDelegateCall_Void |  master |    Core |   2.517 us | 0.0191 us | 0.0179 us | 0.2174 |   1.34 KB |
|           DispatchDelegateCall_Void |      PR |    Core |   1.026 us | 0.0195 us | 0.0216 us | 0.0534 |     376 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|      DispatchInterfaceProxyCall_Int |  master |     Clr |   3.046 us | 0.0201 us | 0.0188 us | 0.3510 |   2.17 KB |
|      DispatchInterfaceProxyCall_Int |      PR |     Clr |   1.137 us | 0.0136 us | 0.0127 us | 0.0572 |     357 B |
|      DispatchInterfaceProxyCall_Int |  master |    Core |   3.372 us | 0.0188 us | 0.0176 us | 0.3128 |   1.93 KB |
|      DispatchInterfaceProxyCall_Int |      PR |    Core |   1.068 us | 0.0214 us | 0.0238 us | 0.0534 |     376 B |
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|     DispatchInterfaceProxyCall_Void |  master |     Clr |   2.598 us | 0.0322 us | 0.0286 us | 0.2365 |   1.47 KB |
|     DispatchInterfaceProxyCall_Void |      PR |     Clr |   1.089 us | 0.0157 us | 0.0147 us | 0.0572 |     357 B |
|     DispatchInterfaceProxyCall_Void |  master |    Core |   2.566 us | 0.0197 us | 0.0184 us | 0.2174 |   1.34 KB |
|     DispatchInterfaceProxyCall_Void |      PR |    Core |   1.010 us | 0.0193 us | 0.0214 us | 0.0534 |     376 B |
//
// ToStringCall
//
| ----------------------------------- | ------- | ------- | ---------- | --------- | --------- | ------ | --------- |
|                    AbstractToString |  master |     Clr | 37.6144 ns | 0.1120 ns | 0.0993 ns | 0.0247 |     104 B |
|                    AbstractToString |      PR |     Clr | 38.2191 ns | 0.1452 ns | 0.1359 ns | 0.0247 |     104 B |
…