@dtchepak 2019-04-11
Thanks a lot for this @zvirja! I will try to work through this on the weekend.

> I'm sane and fully admit that it makes negligible difference in real life, but I still like to see and realize that our tool is improved and fast

I can definitely appreciate this. I think this is a sign of an excellent craftsperson! They take care of the details that others won't always notice from the outside, but are still an important part of a job well done. 👍 😄